backend/requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
tqdm
openai
pytesseract
pdf2image
pdfminer.six
numpy
rich
nltk
python-dotenv
